+++ src/devices/HardwareMonitor/ComputerJson.cs
@@ -1,0 +1,36 @@
+﻿using System;
+
+namespace HardwareMonitor
+{
+    internal class ComputerJson
+    {
+        public ComputerJson()
+        {
+            ComputerName = string.Empty;
+            Hardware = Array.Empty<HardwareJson>();
+        }
+
+        public string ComputerName
+        {
+            get;
+            set;
+        }
+
+        public int LogicalProcessorCount
+        {
+            get;
+            set;
+        }
+
+        public HardwareJson[] Hardware
+        {
+            get;
+            set;
+        }
+
+        public override string ToString()
+        {
+            return ComputerName;
+        }
+    }
+}
